[Tuner] Refactor: Extract rocm_common.py from common.py
