+++ amdsharktuner/amdsharktuner/candidate_ordering.py
@@ -10,6 +10,7 @@ from typing import Optional, Callable
 from iree.compiler.dialects import iree_gpu  # type: ignore
 
 from . import common
+from .rocm import rocm_common
 
 
 class CandidateOrderKind(str, Enum):
@@ -36,7 +37,7 @@ def arith_intensity(x: int, y: int, z: int) -> float:
 
 def llvm_gpu_vector_distribute_contraction_sort_key(
     target_info: iree_gpu.TargetInfo,
-    knob: common.LLVMGPUVectorDistributeContractionKnobs,
+    knob: rocm_common.LLVMGPUVectorDistributeContractionKnobs,
 ) -> tuple[bool, bool, float]:
     return (
         not is_pow2(knob.tile_k),
@@ -50,7 +51,7 @@ def llvm_gpu_vector_distribute_contraction_sort_key(
 
 
 SORT_KEY_MAP: dict[type[common.KnobAssignment | None], Callable | None] = {
-    common.LLVMGPUVectorDistributeContractionKnobs: llvm_gpu_vector_distribute_contraction_sort_key,
+    rocm_common.LLVMGPUVectorDistributeContractionKnobs: llvm_gpu_vector_distribute_contraction_sort_key,
     type(None): None,
     # TODO: Add key() for conv, attention, and other dispatch kinds.
 }
